+++ src/uuid-tool.ts
@@ -1,13 +1,13 @@
-import { UUID } from './uuid';
+import { Uuid } from './uuid';
 
-export class UUIDTool {
+export class UuidTool {
   /**
    * Converts an UUID string to an UUID byte array.
    * @param uuid UUID string e.g. '1FBD384C-B2A1-41C6-84AF-43CABDF44124'
    * @return UUID byte array. (length: 16)
    */
   public static toBytes(uuid: string): number[] {
-    return new UUID().fromString(uuid).toBytes();
+    return new Uuid().fromString(uuid).toBytes();
   }
 
   /**
@@ -16,15 +16,15 @@ export class UUIDTool {
    * @return UUID string. (length 36)
    */
   public static toString(bytes: number[]): string {
-    return new UUID().fromBytes(bytes).toString();
+    return new Uuid().fromBytes(bytes).toString();
   }
 
   /**
    * Generates a new (pseudo) UUID RFC 4122 Version 4.
    * @return UUID e.g. '1FBD384C-B2A1-41C6-84AF-43CABDF44124'
    */
-  public static newUUID(): string {
-    return new UUID().toString();
+  public static newUuid(): string {
+    return new Uuid().toString();
   }
 
   /**
@@ -32,8 +32,8 @@ export class UUIDTool {
    * @param uuid UUID string e.g. '1FBD384C-B2A1-41C6-84AF-43CABDF44124'
    * @return True if valid, otherwise false.
    */
-  public static isUUID(uuid: string): boolean {
-    return new UUID().fromString(uuid).isValid();
+  public static isUuid(uuid: string): boolean {
+    return new Uuid().fromString(uuid).isValid();
   }
 
   /**
@@ -43,6 +43,6 @@ export class UUIDTool {
    * @return True if equal, otherwise false.
    */
   public static compare(uuid1: string, uuid2: string): boolean {
-    return new UUID(uuid1).equals(new UUID(uuid2));
+    return new Uuid(uuid1).equals(new Uuid(uuid2));
   }
 }