+++ src/uuid.ts
@@ -1,11 +1,33 @@
-export class UUID {
-  public static stringExportFormat = 'uppercase';
+const REGEX = {
+  uuid: /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i,
+};
 
-  private static REGEX = {
-    UUID: /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i,
-  };
+export interface UuidLike {
+  id: string;
+}
 
-  private value = '';
+export class Uuid {
+  public static stringExportFormat = 'lowercase';
+
+  /**
+   * Converts the JSON representation of this class to this class.
+   * Example JSON: `{ "id": "...THE UUID..." }`.
+   * Use: `JSON.stringify(thisClassInstance)` to convert this class to JSON string.
+   * @param json Json object or string.
+   */
+  public static fromJson(json: string | UuidLike): Uuid {
+    if (typeof json === 'string') {
+      json = JSON.parse(json) as UuidLike;
+    }
+
+    if (!json || !json.id) {
+      throw new TypeError('The input cannot be converted to Uuid.');
+    }
+
+    return new Uuid(json.id);
+  }
+
+  private id = '';
 
   /**
    * UUID
@@ -25,8 +47,10 @@ export class UUID {
    * Re-generates a new UUID for this instance.
    */
   public generate(): this {
-    // Generate pseudo uuid by random bytes.
-    const bytes = new Array(16).fill(0).map(() => (Math.random() * 0xff) & 0xff);
+    // Generate pseudo UUID by random bytes.
+    const bytes = new Array(16)
+      .fill(0)
+      .map(() => (Math.random() * 0xff) & 0xff);
     // Brand version.
     bytes[6] = (bytes[6] | 0x40) & 0x4f;
     // Parse it as string.
@@ -40,12 +64,12 @@ export class UUID {
    * @param bytes UUID Byte array (length 16 bytes of bytes)
    */
   public fromBytes(bytes: number[]): this {
-    this.value = bytes
+    this.id = bytes
       .map((b) => ('00' + b.toString(16)).slice(-2))
       .join('')
-      .replace(/(.{8})(.{4})(.{4})(.{4})(.{12})/, '$1-$2-$3-$4-$5')
-      .toUpperCase();
+      .replace(/(.{8})(.{4})(.{4})(.{4})(.{12})/, '$1-$2-$3-$4-$5');
 
+    this.id = this.toString();
     return this;
   }
 
@@ -54,7 +78,9 @@ export class UUID {
    * @return UUID byte array. (length: 16)
    */
   public toBytes(): number[] {
-    return (this.value.replace(/-/g, '').match(/.{2}/g) || []).map((b) => parseInt(b, 16));
+    return (this.id.replace(/-/g, '').match(/.{2}/g) || []).map((b) =>
+      parseInt(b, 16)
+    );
   }
 
   /**
@@ -62,7 +88,8 @@ export class UUID {
    * @param uuid UUID string e.g. '1FBD384C-B2A1-41C6-84AF-43CABDF44124'
    */
   public fromString(uuid: string): this {
-    this.value = uuid.trim();
+    this.id = uuid.trim();
+    this.id = this.toString();
     return this;
   }
 
@@ -71,13 +98,12 @@ export class UUID {
    * @return UUID string e.g. '1FBD384C-B2A1-41C6-84AF-43CABDF44124'
    */
   public toString(): string {
-    switch (UUID.stringExportFormat) {
+    switch (Uuid.stringExportFormat) {
       default:
-      case 'uppercase':
-        return this.value.toUpperCase();
-
       case 'lowercase':
-        return this.value.toLowerCase();
+        return this.id.toLowerCase();
+      case 'uppercase':
+        return this.id.toUpperCase();
     }
   }
 
@@ -86,20 +112,20 @@ export class UUID {
    * @return True if valid, otherwise false.
    */
   public isValid(): boolean {
-    if (typeof this.value !== 'string' || this.value.length !== 36) {
+    if (typeof this.id !== 'string' || this.id.length !== 36) {
       return false;
     }
-    return UUID.REGEX.UUID.test(this.value);
+    return REGEX.uuid.test(this.id);
   }
 
   /**
    * Compares this UUID with another UUID.
-   * @param uuid Another UUID or uuid-string.
+   * @param uuid Another UUID or UUID-string.
    * @return True if equal, otherwise false.
    */
-  public equals(uuid: UUID | string): boolean {
+  public equals(uuid: Uuid | string): boolean {
     if (typeof uuid === 'string') {
-      uuid = new UUID(uuid);
+      uuid = new Uuid(uuid);
     }
 
     return this.toString() === uuid.toString();